rename: 메인 클래스와 Racing 클래스 패키지 이동
